Protección de acciones de post
Se completó la política de acceso a las acciones de los post

+++ app/Http/Controllers/Admin/PostController.php
@@ -99,6 +99,8 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        $this->authorize('pass', $post);
+
         return view('admin.posts.show', compact('post'));
     }
 
@@ -128,6 +130,8 @@ class PostController extends Controller
      */
     public function update(UpdatePost $request, Post $post)
     {
+        $this->authorize('pass', $post);
+
         //Llena los campos con la nueva información y luego la guarda
         $post->fill($request->except('file'))->save();
 
@@ -151,6 +155,8 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        $this->authorize('pass', $post);
+        
         $post->delete();
 
         return back()->with('info', 'Entrada ' . $post->name . ' eliminada exitosamente');
